move image already present check to image manager
