+++ pkg/images/image_helpers.go
@@ -17,6 +17,7 @@ limitations under the License.
 package images
 
 import (
+	"encoding/json"
 	"fmt"
 	"strings"
 	"time"
@@ -30,9 +31,9 @@ import (
 )
 
 // newImagePullJob constructs a job manifest for pulling an image to a node
-func newImagePullJob(imagecache *fledgedv1alpha1.ImageCache, image string, hostname string, imagePullPolicy string) (*batchv1.Job, error) {
+func newImagePullJob(imagecache *fledgedv1alpha1.ImageCache, image string, node *corev1.Node, imagePullPolicy string) (*batchv1.Job, error) {
 	var pullPolicy corev1.PullPolicy = corev1.PullIfNotPresent
-
+	hostname := node.Labels["kubernetes.io/hostname"]
 	if imagecache == nil {
 		glog.Error("imagecache pointer is nil")
 		return nil, fmt.Errorf("imagecache pointer is nil")
@@ -131,7 +132,8 @@ func newImagePullJob(imagecache *fledgedv1alpha1.ImageCache, image string, hostn
 }
 
 // newImageDeleteJob constructs a job manifest to delete an image from a node
-func newImageDeleteJob(imagecache *fledgedv1alpha1.ImageCache, image string, hostname string, containerRuntimeVersion string, dockerclientimage string) (*batchv1.Job, error) {
+func newImageDeleteJob(imagecache *fledgedv1alpha1.ImageCache, image string, node *corev1.Node, containerRuntimeVersion string, dockerclientimage string) (*batchv1.Job, error) {
+	hostname := node.Labels["kubernetes.io/hostname"]
 	if imagecache == nil {
 		glog.Error("imagecache pointer is nil")
 		return nil, fmt.Errorf("imagecache pointer is nil")
@@ -223,4 +225,32 @@ func newImageDeleteJob(imagecache *fledgedv1alpha1.ImageCache, image string, hos
 		job.Spec.Template.Spec.Volumes[0].VolumeSource.HostPath.Path = "/var/run/crio/crio.sock"
 	}
 	return job, nil
+}
+
+func checkIfImageNeedsToBePulled(imagePullPolicy string, image string, node *corev1.Node) (bool, error) {
+	var fImage string
+	if imagePullPolicy == string(corev1.PullIfNotPresent) {
+		if !strings.Contains(image, ":") && !strings.Contains(image, "@sha") {
+			fImage = strings.Join([]string{image, "latest"}, ":")
+		}
+		imageAlreadyPresent, err := imageAlreadyPresentInNode(fImage, node)
+		if err != nil {
+			return false, err
+		}
+		if imageAlreadyPresent {
+			return false, nil
+		}
+	}
+	return true, nil
+}
+
+func imageAlreadyPresentInNode(image string, node *corev1.Node) (bool, error) {
+	imagesByteSlice, err := json.Marshal(node.Status.Images)
+	if err != nil {
+		return false, err
+	}
+	if strings.Contains(string(imagesByteSlice), image) {
+		return true, nil
+	}
+	return false, nil
 }
